+++ src/aquiche/__init__.py
@@ -11,14 +11,14 @@ from aquiche._expiration import CacheExpirationValue, DurationExpirationValue
 from aquiche._hash import Key
 
 __all__ = [
-    AquicheFunctionWrapper,
-    CacheInfo,
-    CacheParameters,
-    alru_cache,
-    clear_all,
-    clear_all_sync,
-    CachedItem,
-    CacheExpirationValue,
-    DurationExpirationValue,
-    Key,
+    "AquicheFunctionWrapper",
+    "CacheInfo",
+    "CacheParameters",
+    "alru_cache",
+    "clear_all",
+    "clear_all_sync",
+    "CachedItem",
+    "CacheExpirationValue",
+    "DurationExpirationValue",
+    "Key",
 ]