+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Simplified the interface
 - Renamed `wrap_async_exit_stack` to `async_context`
 - Added `refresh_threshold` param
+- Removed `exit_stack_close_delay` cache parameter
 - Removed `await_exit_stack_close_operations` and `cancel_exit_stack_close_operations` functions
 - Updated documentation to reflect new changes
 - Changed build system from poetry to uv